@@ -1,12 +1,17 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
   <PropertyGroup>
     <OutputType>Exe</OutputType>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <Version>1.5.3</Version>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <Version>1.5.4</Version>
     <LangVersion>latest</LangVersion>
+    <AspNetCoreHostingModel>OutOfProcess</AspNetCoreHostingModel>
+    <AspNetCoreModuleName>AspNetCoreModuleV2</AspNetCoreModuleName>
     <AssemblyName>Lykke.Service.CandlesHistory</AssemblyName>
     <RootNamespace>Lykke.Service.CandlesHistory</RootNamespace>
   </PropertyGroup>
+  <ItemGroup>
+    <FrameworkReference Include="Microsoft.AspNetCore.App" />
+  </ItemGroup>
   <PropertyGroup>
     <NoWarn>1701;1702;1705;CA2007</NoWarn>
     <UserSecretsId>F7428201-B7F7-4571-9ABD-3A1F4CF87613</UserSecretsId>
