@@ -5,7 +5,7 @@ using Common.Log;
 using Lykke.Service.CandlesHistory.Core.Domain.Candles;
 using System;
 using System.Collections.Generic;
-using System.Data.SqlClient;
+using Microsoft.Data.SqlClient;
 using System.Data.SqlTypes;
 using System.Linq;
 using System.Text;
